move Game/* contents and remove it
