+++ resource/dead.png
